refactor : UserInteraction 타입 새로 구현
- 기존 Bank 구조체에 화면에 표시하는 메써드를 userInteraction 타입으로 이동
