+++ BankManagerConsoleApp/BankManagerConsoleApp/BankMessage.swift
@@ -15,9 +15,4 @@ enum BankMessage {
         2: 종료
         입력:
         """
-    static func bankClosed(customers: Int) -> String {
-        let totalWorkTime = 0.7 * Double(customers)
-        return "업무가 마감되었습니다. 오늘 업무를 처리한 고객은 총 \(customers)명이며, 총 업무시간은 \(totalWorkTime)초 입니다."
-    }
-    
 }
